# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: RDK Robot Learning Platform API
  version: 1.0.0
  description: >-
    Versioned control-plane contract for the standalone MicroDuck/RDK Duck
    Sim2Real workflow. The legacy /api/sim2real paths are aliases.
servers:
  - url: https://example.invalid
    description: Replace with the deployment origin.
...
security:
  - cookieAuth: []
paths:
...
  /api/sim2real/agent/plan:
    post:
      tags: [agent]
      summary: Parse a Chinese task into a guarded Sim2Real tool plan
      operationId: createSim2RealAgentPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1, maxLength: 2000 }
                context: { type: object, additionalProperties: true }
      responses:
        "200": { description: Agent plan. }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/sim2real/agent/execute:
    post:
      tags: [agent]
      summary: Start an asynchronous guarded Agent run
      operationId: executeSim2RealAgentPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan: { type: object, additionalProperties: true }
                approved: { type: boolean, default: false }
      responses:
        "202": { description: Run accepted. }
        "409": { description: Explicit approval required. }
  /api/sim2real/agent/runs/{runId}:
    get:
      tags: [agent]
      summary: Read Agent run status, tool events and evidence
      operationId: getSim2RealAgentRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200": { description: Agent run snapshot. }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: rdk_sso_web_session
      description: >-
        Site-wide encrypted session cookie issued by the Studio main shell;
        the reference studio-cookie-auth adapter verifies it with
        AES-256-GCM (v1.iv.tag.ciphertext). The reference local adapter has
        no login.
  parameters:
...
    RunId:
      in: path
      name: runId
      required: true
      schema: { type: string }
...
  responses:
...
    BadRequest:
      description: Input or contract validation failed.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource is not visible to this account.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { const: false }
        error: { type: string }
        code: { type: string }
        message: { type: string, description: Human-readable detail when available. }
        requestId: { type: string }
        retryable: { type: boolean }
...